[CHANGED] topo_id use he::TopoID

+++ include/everything/GeoAttrRebuild.h
@@ -24,10 +24,10 @@ private:
 private:
     GeometryImpl& m_geo;
 
-    std::map<int, std::vector<VarValue>> m_points;
-    std::map<int, std::vector<VarValue>> m_vertices;
-    std::map<int, std::vector<VarValue>> m_primitives;
-    std::vector<VarValue>                m_detail;
+    std::map<uint64_t, std::vector<VarValue>> m_points;
+    std::map<uint64_t, std::vector<VarValue>> m_vertices;
+    std::map<uint64_t, std::vector<VarValue>> m_primitives;
+    std::vector<VarValue>                     m_detail;
 
     std::vector<GeoAttribute::VarDesc> m_var_descs[static_cast<int>(GeoAttrType::MaxTypeNum)];
